Cholesky kinda works, it only gives trivial solution
Setting a coefficient to a value different than zero makes all of the
other values different than zero, the problem is that without doing that
by default it goes to the trivial case.

The problem with setting a coefficient to a value different than zero is
that it becomes a very hard problem trying to set it to a value that makes it
numerically stable.

Right now trying to find a way to make it possible.